fix: naming, docs, tests, throws when MIME type fails
