+++ addons/with-web/data-url.js
@@ -7,6 +7,10 @@ module.exports = {
     process(src, filename) {
         const mimeType = mime.getType(filename);
 
+        if (mimeType == null) {
+            throw new Error('Could not guess MIME type of file.');
+        }
+
         return `module.exports = ${JSON.stringify(`data:${mimeType};base64,${fs.readFileSync(filename).toString('base64')}`)};`;
     },
 };
