+++ composer.js
@@ -1,19 +1,15 @@
 'use strict';
 
 module.exports = (listOfAddons = []) => {
-    let aggregator = {};
-
     if (!Array.isArray(listOfAddons)) {
         throw new Error('Received argument is not an array.');
     }
 
-    for (const addon of listOfAddons) {
+    return listOfAddons.reduce((acc, addon) => {
         if (typeof addon !== 'function') {
             throw new Error('Found item in argument list that is not a function, all objects to compose should be functions!');
         }
 
-        aggregator = addon(aggregator);
-    }
-
-    return aggregator;
+        return addon(acc);
+    }, {});
 };
