--- test/jestConfig.test.js
@@ -1,7 +1,0 @@
-'use strict';
-
-const jestConfig = require('../jestConfig');
-
-it('should match snapshot', () => {
-    expect(jestConfig()).toMatchSnapshot();
-});
