+++ tests/baseConfig.test.js
@@ -1,0 +1,7 @@
+'use strict';
+
+const baseConfig = require('../baseConfig');
+
+it('should match snapshot', () => {
+    expect(baseConfig()).toMatchSnapshot();
+});
