+++ tests/with-web.test.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const configWeb = require('../addons/withWeb');
+const configWeb = require('../addons/with-web');
 
 it('should match snapshot', () => {
     expect(configWeb()).toMatchSnapshot();
